requirements.txt
Django==6.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
psycopg2-binary==2.9.11
pytest==8.3.4
pytest-django==4.9.0